+++ index.html
@@ -33,13 +33,15 @@
             </div>
             <ul id="calculators" class="sidebar-section">
                 <li>
-                    <a href="./tools/modsCalculator.html" class="nav-link" title="Mods Calculator" data-tooltip="Mods Calculator">
+                    <a href="./tools/modsCalculator.html" class="nav-link" title="Mods Calculator"
+                        data-tooltip="Mods Calculator">
                         <span class="icon">📊</span>
                         <span class="label">Modifiers</span>
                     </a>
                 </li>
                 <li>
-                    <a href="./tools/prestigeCalculator.html" class="nav-link" title="Prestige Calculator" data-tooltip="Prestige Calculator">
+                    <a href="./tools/prestigeCalculator.html" class="nav-link" title="Prestige Calculator"
+                        data-tooltip="Prestige Calculator">
                         <span class="icon">💎</span>
                         <span class="label">Prestige</span>
                     </a>
@@ -71,9 +73,9 @@
             <div class="changelog">
                 <p>Version 0.1 :
                     <br>
-                    <span>- Modifiers calculator functionnal</span>
+                    <span>- Modifiers calculator functional</span>
                     <br>
-                    <span>- Prestige calculator functionnal</span>
+                    <span>- Prestige calculator functional</span>
                     <br>
                     <span>- Responsive of the website</span>
                     <br>
