+++ tools/modsCalculator.html
@@ -117,7 +117,7 @@
                 </div>
 
                 <div id="crafterModExtras">
-                    
+
                 </div>
             </div>
         </div>
